backend/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "zazzle-backend"
version = "0.1.0"
description = "Uzbekistan Print-on-Demand Platform Backend"
authors = [
    {name = "Zazzle Team", email = "dev@zazzle.uz"}
]
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.11"
dependencies = [
    # Django Core
    "Django>=4.2,<5.0",
    "djangorestframework>=3.14.0",
    "django-cors-headers>=4.3.0",
    "django-filter>=23.3",
    "dj-rest-auth>=5.0.0",
    "django-allauth>=0.57.0",
    
    # Database & Cache
    "psycopg2-binary>=2.9.7",
    "redis>=4.6.0",
    "django-redis>=5.4.0",
    
    # Celery for async tasks
    "celery>=5.3.0",
    "django-celery-beat>=2.5.0",
    "django-celery-results>=2.5.0",
    
    # AWS & File Storage
    "boto3>=1.28.0",
    "django-storages>=1.14.0",
    "Pillow>=10.0.0",
    
    # API & Serialization
    "djangorestframework-simplejwt>=5.3.0",
    "drf-spectacular>=0.26.0",
    "django-extensions>=3.2.0",
    
    # Environment & Configuration
    "python-decouple>=3.8",
    "dj-database-url>=2.1.0",
    
    # Payment Processing
    "stripe>=6.6.0",
    
    # Utils
    "requests>=2.31.0",
    "python-slugify>=8.0.0",
    "Faker>=19.6.0",
]

[project.optional-dependencies]
dev = [
    # Development Dependencies
    "django-debug-toolbar>=4.2.0",
    "ipython>=8.15.0",
    "django-extensions>=3.2.0",
    
    # Testing
    "pytest>=7.4.0",
    "pytest-django>=4.5.0",
    "pytest-cov>=4.1.0",
    "factory-boy>=3.3.0",
    "model-bakery>=1.12.0",
    
    # Code Quality
    "black>=23.7.0",
    "ruff>=0.0.291",
    "mypy>=1.5.0",
    "django-stubs>=4.2.0",
    "djangorestframework-stubs>=3.14.0",
    "types-requests>=2.31.0",
    "types-redis>=4.6.0",
    "pre-commit>=3.4.0",
]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | migrations
)/
'''

[tool.ruff]
target-version = "py311"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"settings.py" = ["F405", "F403"]
"*/migrations/*.py" = ["E501"]

[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]

[tool.django-stubs]
django_settings_module = "zazzle.settings"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "zazzle.settings.test"
python_files = ["test_*.py", "*_test.py", "testing/***/*.py"]

[tool.coverage.run]
source = "."
omit = [
    "*/venv/*",
    "*/migrations/*",
    "*/settings/*",
    "manage.py",
    "*/tests/*",
    "*/__init__.py"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]